requirements.txt
streamlit>=1.35
pandas>=2.0
httpx>=0.27
plotly>=5.22
kaleido>=0.2
scikit-learn>=1.4
tensorflow>=2.16
anthropic>=0.30
python-dotenv>=1.0
